Generate versioned documentation
In the past, we only had one version of the documentation online. This
caused confusion as to which OpTiMSoC version a documentation
references. Due to the frequent experiments with the installation of the
individual tools, using documentation with a wrong version usually
fails.

To make getting started with OpTiMSoC easier, we now build documentation
for each OpTiMSoC release separately. We list all available versions
online.