Build refman with Sphinx
If the refman is present ("index.rst") build it using Sphinx. Sphinx
is installed locally in a virtual environment.

We supply a simple "theme" which actually generates liquid templates
to be further processed by Jekyll.

The Sphinx build is prepared in a way that it contains the user guide
and reference manual (and potentially other documentation) as parts,
containing them in a unified tree. This is helpful if we in any case
want to relocate it and orients at the Open SoC Debug solution.